refactor(cli): hook commands move to src/cli/hooks.ts
Move codex-hook, restore, session-end, user-prompt, post-tool and
session-snapshot out of bin/lcm.ts's main() into
registerHookCommands(program) in src/cli/hooks.ts, called at the
position the block occupied (immediately before the mcp command).

readStdin moves into src/cli/support.ts since both the hook commands
and the compact block (which stays in bin/lcm.ts) use it.
